+++ src/main/kotlin/dev/ekvedaras/laravelquery/models/DbReferenceExpression.kt
@@ -6,14 +6,21 @@ import com.intellij.database.model.DasIndex
 import com.intellij.database.model.DasNamespace
 import com.intellij.database.model.DasTable
 import com.intellij.database.model.DasTableKey
+import com.intellij.openapi.Disposable
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.application.ReadAction
 import com.intellij.openapi.project.DumbService
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.Disposer
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiManager
+import com.intellij.psi.PsiTreeChangeAdapter
+import com.intellij.psi.PsiTreeChangeEvent
 import dev.ekvedaras.laravelquery.utils.DbReferenceResolver
 import dev.ekvedaras.laravelquery.utils.PsiUtils.Companion.unquoteAndCleanup
 import dev.ekvedaras.laravelquery.utils.TableAndAliasCollector
+import kotlin.math.exp
 import org.apache.commons.lang.StringUtils.substringBefore
 
 class DbReferenceExpression(val expression: PsiElement, val type: Type) {
@@ -55,11 +62,17 @@ class DbReferenceExpression(val expression: PsiElement, val type: Type) {
             ranges.add(TextRange.from(if (ranges.isNotEmpty()) ranges.last().endOffset + 1 else 1, part.length))
         }
 
-        if (!DumbService.isDumb(project) && ApplicationManager.getApplication().isReadAccessAllowed) {
-            ApplicationManager.getApplication().runReadAction {
+        if (!DumbService.isDumb(project)) {
+            val expressionDisposable = Disposer.newDisposable()
+            PsiManager.getInstance(project).addPsiTreeChangeListener(object : PsiTreeChangeAdapter() {
+                override fun childrenChanged(event: PsiTreeChangeEvent) {
+                    expressionDisposable.dispose()
+                }
+            }, expressionDisposable)
+            ReadAction.nonBlocking<Unit> {
                 TableAndAliasCollector(this).collect()
                 DbReferenceResolver(this).resolve()
-            }
+            }.inSmartMode(project).expireWith(expressionDisposable).executeSynchronously()
         }
     }
 }